add basic bruteforce change to collinear.

+++ collinear/Point.java
@@ -21,6 +21,7 @@ public class Point implements Comparable<Point> {
      *
      * @param  x the <em>x</em>-coordinate of the point
      * @param  y the <em>y</em>-coordinate of the point
+     * assume that the x and y are each between 0 and 32,767
      */
     public Point(int x, int y) {
         /* DO NOT MODIFY */
@@ -66,7 +67,7 @@ public class Point implements Comparable<Point> {
         } else if (this.y == that.y) {
             return 0.0;
         } else {
-            return (that.y - this.y) / (that.x - this.x);
+            return (double) (that.y - this.y) / (that.x - this.x);
         }
     }
